Moved tests from editing, and added new to test initiall state.
