+++ packages/ckeditor5-find-and-replace/tests/findandreplaceediting.js
@@ -103,63 +103,6 @@ describe( 'FindAndReplaceEditing', () => {
 		} );
 	} );
 
-	describe( 'highlighted result handling', () => {
-		it( 'adds a highlight marker', () => {
-			editor.setData( FOO_BAR_PARAGRAPH );
-
-			const paragraph = root.getChild( 0 );
-			const marker = addMarker( 'findResult:test-uid', paragraph, 4, 7 );
-
-			const matchInfo = addSearchResultToState( marker );
-
-			editor.plugins.get( 'FindAndReplaceEditing' ).state.highlightedResult = matchInfo;
-
-			expect( editor.model.markers.has( 'findResultHighlighted:test-uid' ) ).to.be.true;
-		} );
-
-		it( 'removes the previous highlight marker', () => {
-			editor.setData( FOO_BAR_PARAGRAPH );
-
-			const paragraph = root.getChild( 0 );
-			const marker = addMarker( 'findResult:test-uid', paragraph, 4, 7 );
-			const highlightedMatch = addSearchResultToState( marker );
-			editor.plugins.get( 'FindAndReplaceEditing' ).state.highlightedResult = highlightedMatch;
-
-			editor.plugins.get( 'FindAndReplaceEditing' ).state.highlightedResult = null;
-
-			expect( editor.model.markers.has( 'findResultHighlighted:test-uid' ) ).to.be.false;
-		} );
-
-		it( 'moves the highlight marker', () => {
-			editor.setData( FOO_BAR_PARAGRAPH );
-
-			const paragraph = root.getChild( 0 );
-			const firstMarker = addMarker( 'findResult:test1', paragraph, 1, 3 );
-			const secondMarker = addMarker( 'findResult:test2', paragraph, 4, 6 );
-			const firstMatch = addSearchResultToState( firstMarker );
-			const secondMatch = addSearchResultToState( secondMarker );
-
-			editor.plugins.get( 'FindAndReplaceEditing' ).state.highlightedResult = firstMatch;
-			editor.plugins.get( 'FindAndReplaceEditing' ).state.highlightedResult = secondMatch;
-
-			expect( editor.model.markers.has( 'findResultHighlighted:test1' ) ).to.be.false;
-			expect( editor.model.markers.has( 'findResultHighlighted:test2' ) ).to.be.true;
-		} );
-
-		function addSearchResultToState( marker ) {
-			const state = editor.plugins.get( 'FindAndReplaceEditing' ).state;
-			const matchInfo = {
-				id: marker.name.replace( /^findResult:/, '' ),
-				label: 'label',
-				marker
-			};
-
-			state.results.add( matchInfo );
-
-			return matchInfo;
-		}
-	} );
-
 	describe( 'commands', () => {
 		it( 'should register find command', () => {
 			expect( editor.commands.get( 'find' ) ).to.be.instanceOf( FindCommand );
@@ -171,64 +114,6 @@ describe( 'FindAndReplaceEditing', () => {
 
 		it( 'should register replace all command', () => {
 			expect( editor.commands.get( 'replaceAll' ) ).to.be.instanceOf( ReplaceAllCommand );
-		} );
-	} );
-
-	describe( 'state', () => {
-		it.skip( 'should automatically remove unused marker', () => {
-			const state = editor.plugins.get( 'FindAndReplaceEditing' ).state;
-
-			editor.setData( '<p>foo foo foo</p>' );
-
-			editor.execute( 'find', 'foo' );
-
-			state.results.remove( 1 );
-
-			const markers = Array.from( editor.model.markers ).filter( markers => markers.name.startsWith( 'findResult:' ) );
-
-			expect( markers ).to.have.length( 2 );
-		} );
-
-		describe( 'changing highlighted result', () => {
-			it( 'should automatically change highlighted result', () => {
-				const state = editor.plugins.get( 'FindAndReplaceEditing' ).state;
-
-				editor.setData( '<p>foo foo foo</p>' );
-				editor.execute( 'find', 'foo' );
-
-				const expectedHighlightedResult = state.results.get( 1 );
-
-				state.results.remove( 0 );
-
-				expect( state.highlightedResult ).to.eql( expectedHighlightedResult );
-			} );
-
-			it( 'should automatically change last highlighted result', () => {
-				const state = editor.plugins.get( 'FindAndReplaceEditing' ).state;
-
-				editor.setData( '<p>foo foo foo</p>' );
-				editor.execute( 'find', 'foo' );
-				editor.execute( 'findNext' );
-				editor.execute( 'findNext' );
-
-				const expectedHighlightedResult = state.results.get( 0 );
-
-				state.results.remove( 2 );
-
-				expect( state.highlightedResult ).to.eql( expectedHighlightedResult );
-			} );
-
-			it( 'should remove highlighted result if there is nothing more to highlight', () => {
-				const state = editor.plugins.get( 'FindAndReplaceEditing' ).state;
-
-				editor.setData( '<p>foo </p>' );
-
-				editor.execute( 'find', 'foo' );
-
-				state.results.remove( 0 );
-
-				expect( state.highlightedResult ).to.be.null;
-			} );
 		} );
 	} );
